+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added detailed comment-based help to all public functions in the `WisherTools.Helpers` module. This provides better guidance and documentation for the following functions:
+- Added detailed comment-based help and Unit Tests to all public functions in the `WisherTools.Helpers` module. This provides better guidance and documentation for the following functions:
   - `Get-DirectoryPath`
   - `Get-FunctionScriptBlock`
   - `New-UniqueFilePath`